requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
black==25.1.0
certifi==2025.4.26
click==8.2.0
colorama==0.4.6
distro==1.9.0
Django==5.2.1
flake8==7.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.9.0
mccabe==0.7.0
mypy_extensions==1.1.0
openai==1.78.1
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pycodestyle==2.13.0
pydantic==2.11.4
pydantic_core==2.33.2
pyflakes==3.3.2
python-dotenv==1.1.0
PyYAML==6.0.2
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
